requirements.txt
lightning==2.3.0
hydra-core==1.3
hydra-zen
hydra-colorlog
rich

numpy==1.26.4
jupyter
matplotlib
ipdb
setuptools>=68.0
black
tensorboardX
opencv-python
ffmpeg-python
scikit-image
termcolor
einops
imageio==2.34.1
av==13.0.0  # imageio[pyav], improved performance over imageio[ffmpeg]
joblib

trimesh
chumpy
smplx
wis3d
pycolmap
open3d==0.17.0

ultralytics==8.2.42  # YOLO
cython_bbox
lapx
transformers==4.38.0
pillow==10.2.0
huggingface-hub==0.26.5
onnxruntime
rembg

absl-py==2.3.1
addict==2.4.0
easydict==1.13
loguru==0.7.3

omegaconf==2.3.0
hydra-submitit-launcher==1.2.0
submitit==1.5.3

tqdm==4.67.1
pandas==2.3.3

pycocotools==2.0.10
yacs==0.1.8
fvcore==0.1.5.post20221221
iopath==0.1.10
detectron2==0.6

plyfile==1.1.3
transforms3d==0.4.2
pyquaternion==0.9.9
pykalman==0.11.0
xatlas==0.0.11
pymeshfix==0.17.2
pyrender==0.1.45
pyopengl==3.1.0
pyglet==2.1.11
pyvista==0.46.4
vtk==9.5.2

